requirements.txt
# Flask Framework
Flask==3.1.1
Werkzeug==3.1.3

# Flask Extensions
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-CORS==6.0.0
Flask-WTF==1.1.1
Flask-Limiter==3.5.0

# Database
SQLAlchemy==2.0.41
psycopg2-binary==2.9.9

# Templates and Utilities
Jinja2==3.1.6
MarkupSafe==3.0.2
WTForms==3.0.1
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0

# HTTP Requests
requests==2.31.0

# Environment
python-dotenv==1.0.0

# System Dependencies
blinker==1.9.0
click==8.2.1
itsdangerous==2.2.0
psutil==6.1.0
typing_extensions==4.14.0

# AI and External APIs
openai==1.3.5

# Date/Time utilities
python-dateutil==2.8.2

# Scheduling
APScheduler==3.10.4

# Email support (optional)
Flask-Mail==0.9.1

# Redis for caching
redis==5.0.1
hiredis==2.2.3

# Testing (development)
pytest==7.4.3
pytest-flask==1.3.0

# Production server
gunicorn==21.2.0

# Development and Testing (optional)
pytest-cov==4.1.0
